requirements.txt
asn1crypto==0.24.0
cffi==1.12.1
Click==7.0
cryptography==2.5
Flask==1.0.2
Flask-Inputs==0.3.0
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
pymssql==2.1.4
six==1.12.0
Werkzeug==0.14.1
WTForms==2.2.1
